+++ src/assetbundles/retour/src/vue/DashboardChart.vue
@@ -48,7 +48,11 @@
             siteId: {
                 type: Number,
                 default: 0,
-            }
+            },
+            refreshIntervalSecs: {
+                type: Number,
+                default: 3,
+            },
         },
         methods: {
             // Load in our chart data asynchronously
@@ -74,10 +78,12 @@
             this.getSeriesData();
         },
         mounted() {
-            // Live refresh the data
-            setInterval(() => {
-                this.getSeriesData();
-            }, 3000);
+            // Live refresh the data -- disabled, causes more problems then helps
+            if (this.refreshIntervalSecs) {
+                setInterval(() => {
+                    this.getSeriesData();
+                }, this.refreshIntervalSecs * 1000);
+            }
         },
         data: function() {
             return {