+++ src/assetbundles/retour/src/vue/DashboardTable.vue
@@ -49,7 +49,11 @@
             siteId: {
                 type: Number,
                 default: 0,
-            }
+            },
+            refreshIntervalSecs: {
+                type: Number,
+                default: 3,
+            },
         },
         data: function() {
             return {
@@ -75,13 +79,15 @@
             this.$events.$on('filter-set', eventData => this.onFilterSet(eventData));
             this.$events.$on('filter-reset', e => this.onFilterReset());
             // Live refresh the data
-            setInterval(() => {
-                if ((typeof this.$refs.pagination !== 'undefined') && (this.$refs.pagination.isOnFirstPage)) {
-                    if (typeof this.$refs.vuetable !== 'undefined') {
-                        this.$refs.vuetable.refresh();
+            if (this.refreshIntervalSecs) {
+                setInterval(() => {
+                    if ((typeof this.$refs.pagination !== 'undefined') && (this.$refs.pagination.isOnFirstPage)) {
+                        if (typeof this.$refs.vuetable !== 'undefined') {
+                            this.$refs.vuetable.refresh();
+                        }
                     }
-                }
-            }, 3000);
+                }, this.refreshIntervalSecs * 1000);
+            }
         },
         methods: {
             onFilterSet (filterText) {