+++ src/models/Settings.php
@@ -11,7 +11,10 @@
 
 namespace nystudio107\retour\models;
 
+use nystudio107\retour\Retour;
+
 use craft\base\Model;
+use craft\behaviors\EnvAttributeParserBehavior;
 use craft\validators\ArrayValidator;
 
 use yii\behaviors\AttributeTypecastBehavior;
@@ -72,6 +75,11 @@ class Settings extends Model
     public $statsStoredLimit = 1000;
 
     /**
+     * @var int Dashboard data live refresh interval
+     */
+    public $refreshIntervalSecs = 5;
+
+    /**
      * @var bool Whether the Statistics should be trimmed after each new statistic is recorded
      */
     public $automaticallyTrimStatistics = true;
@@ -114,6 +122,8 @@ class Settings extends Model
             ['dynamicRedirectDisplayLimit', 'default', 'value' => 100],
             ['statsStoredLimit', 'integer', 'min' => 1],
             ['statsStoredLimit', 'default', 'value' => 1000],
+            ['refreshIntervalSecs', 'integer', 'min' => 0],
+            ['refreshIntervalSecs', 'default', 'value' => 3],
             ['statsDisplayLimit', 'integer', 'min' => 1],
             ['statsDisplayLimit', 'default', 'value' => 1000],
             ['excludePatterns', ArrayValidator::class],
@@ -125,11 +135,22 @@ class Settings extends Model
      */
     public function behaviors()
     {
-        return [
+        $craft31Behaviors = [];
+        if (Retour::$craft31) {
+            $craft31Behaviors = [
+                'parser' => [
+                    'class' => EnvAttributeParserBehavior::class,
+                    'attributes' => [
+                    ],
+                ]
+            ];
+        }
+
+        return array_merge($craft31Behaviors, [
             'typecast' => [
                 'class' => AttributeTypecastBehavior::class,
                 // 'attributeTypes' will be composed automatically according to `rules()`
             ],
-        ];
+        ]);
     }
 }